Part 1 of breaking distributor -> querier package coupling: move most types to `pkg/ingester/client`.
